feat: fixed abac solution for owners and self, added persisted volumes for postgres

+++ server/src/app/modules/books/books.service.ts
@@ -5,7 +5,6 @@ import {
   ResponseBookDto,
   UpdateBookDto,
 } from './dtos/book.dto';
-import ExtractorService from '../../providers/extractor.service';
 import {
   DefaultNotFoundException,
   CollectionNotFoundException,
@@ -13,19 +12,25 @@ import {
 import { InjectRepository } from '@nestjs/typeorm';
 import BookEntity from './entities/book.entity';
 import { DeleteResult, Repository } from 'typeorm';
+import UserEntity from '../users/entities/user.entity';
 
 @Injectable()
 export default class BooksService {
   constructor(
-    private extractor: ExtractorService,
     @InjectRepository(BookEntity)
     private booksRepository: Repository<BookEntity>,
   ) {}
 
   // CRUD //
 
-  async create(bookData: CreateBookDto): Promise<ResponseBookDto> {
-    const newBook = await this.booksRepository.create(bookData);
+  async create(
+    bookData: CreateBookDto,
+    user: UserEntity,
+  ): Promise<ResponseBookDto> {
+    const newBook = await this.booksRepository.create({
+      ...bookData,
+      owner: user,
+    });
     const savedBook = await this.booksRepository.save(newBook);
 
     if (savedBook) return new ResponseBookDto(savedBook);
@@ -49,31 +54,23 @@ export default class BooksService {
   async updateOrPatch(
     id: string,
     bookData: UpdateBookDto | PatchBookDto,
-    requesterId: string,
+    ownerId: string,
   ): Promise<ResponseBookDto> {
     const modifiedBook = await this.booksRepository.create({
       id,
       ...bookData,
-      lastChangedBy: requesterId,
+      lastChangedBy: ownerId,
     });
-    const savedBook = await this.booksRepository.save(modifiedBook);
+    const savedBook = await this.booksRepository.save({
+      ...modifiedBook,
+      ownerId,
+    });
 
     if (savedBook) return new ResponseBookDto(savedBook);
     throw new DefaultNotFoundException(id, 'Book');
   }
 
-  async delete(id: string): Promise<DeleteResult> {
-    return this.booksRepository.delete({ id });
-  }
-
-  // CUSTOM //
-
-  async sync(): Promise<string> {
-    const extractedMsg = this.extractor.sync('../library', [
-      'FearAgentFolders.cbz',
-      '2000AD-2268.cbr',
-    ]);
-
-    return extractedMsg;
+  async delete(id: string, ownerId: string): Promise<DeleteResult> {
+    return this.booksRepository.delete({ id, ownerId });
   }
 }
